added wrappers for HYPRE's BoomerAMG, ParaSails and Euclid preconditioners
This commit adds three preconditioner wrappers, which are implemented in the
HYPRE third-party libary, namely: BoomerAMG, ParaSails and Euclid.
Furthermore, this commit also adds a common Solver::HypreBase helper base-
class for the three former classes, which itself derives from the
Solver::ADPSolverBase base-class, to take care of the conversion of the
matrix and vectors to the corresponding HYPRE objects.

The short description of the three new preconditioners:

* BommerAMG is a parallel algebraic multigrid implementation.
* ParaSails is some sort of parallel sparse approximate inverse preconditioner.
* Euclid is some sort of parallel incomplete LU decomposition preconditioner.

Please note that currently all three preconditioners are created with default
parameters. Unfortunately, I currently do not have the time to fiddle around
with the not-so-useful documention of the HYPRE library to figure out which
of the parameters have to be configured in what manner. So whoever is in the
need of using one of those preconditioners will have to deal with the HYPRE
documentation first and add the corresponding parameter setters to our wrapper
classes.

This commit also adds a small test, which will ensure that all of the above
three preconditioners can be used.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>