use new PMDCDSCMatrix in Navier-Stokes PPnD application
This commit replaces the Global::SymmetricLumpedSchurMatrix by the newly added
Global::PMDCDSCMatrix in the also newly added PPnD application. In addition to
that, a few minor modifications had to be made to make this work correctly:

* The 'lumped_mass_velo' vector had to be replaced by its component-wise
  inverse named 'inverse_lumped_mass_velo'.
* Furthermore, the inverse lumped velocity mass matrix vector is now filtered,
  because this is required by the PMDCDSCMatrix class to work correctly.
* The assembly of the velocity mass matrix is now performed using the
  standard bilinear operator assembly insteady of the BurgersAssembler.
* The 'lumped_mass_pres' vector has been renamed to 'inverse_lumped_mass_pres'
  for the sake of consistency. Note that the vector already contained the
  inverse, so this is a purely syntactical change.

Surprisingly, the switch to the PMDCDSCMatrix has lead to a significant
decrease of Solver-S iterations (~24%) and thus to a significant decrease of
Solver-S runtime (~43%). The exact reasons remain unknown, but we'll happily
accept this gift :)

Furthermore, the default mesh for the "flow-around-a-cylinder" benchmarks
has been changed from the 32-quad mesh "flowbench_c2d_01_quad_32.xml" to the
64-quad mesh "flowbench_c2d_03_quad_64.xml". The latter one has much nicer
geometrical quality around the obstacle than the former one, which may
significantly improve numerical convergence at only twice the amount of DOFs.
This may also allow for larger timesteps, so that the doubled DOF count may
possibly compensated by doubled time step size.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>