outsourced HYPRE wrapper code to separate source file
This commit adds a new source file, which now contains all data structures
and function calls related to the HYPRE third-party library, and provides an
opaque oldschool C-style void*-based wrapper interface.

This approach has been chosen, because HYPRE is a quite large library, which
defines tons of own data types and functions, which, unfortunately, are mostly
included by the two important header files, which leads to massive namespace
pollution in FEAT. Most disturbing is the fact, that HYPRE defines its own
"MPI_Comm" typedef if configured without MPI, thus potentially leading to
conflicts with other libraries. Therefore, the newly introduced "hypre.cpp"
source file is the only file in FEAT, which directly includes any of the
HYPRE header files.

As a positive consequence, it is now possible to use our HYPRE-based solvers
even in non-MPI builds. However, it is still necessary to use the
AlgDofPartiMatrix or the PMDCDSCMatrix interface to do so.

This commit also removes the obsolete "dbg-hypre" tool.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>