added support for SuperLU_DIST thirdparty library
This commit adds the SuperLU_DIST library, which offers an MPI-parallel
sparse direct solver, which is wrapped by the new Solver::SuperLU class,
which uses our algebraic dof partitioning functionality.

The new library can be enabled by adding the 'superlu' tag into the buildid,
however, it is to be noted that SuperLU can only be used in MPI builds,
because we do not include the sequential SuperLU library for non-MPI builds.

Currently, the SuperLU solver only works for matrices and filters of the type
* Global::Matrix<LAFEM::SparseMatrixCSR<Mem::Main, double, *>>
* Global::Filter<LAFEM::UnitFilter<Mem::Main, double, **>
* Global::Filter<LAFEM::NoneFilter<Mem::Main, double, **>

We will add a new Solver::DirectSolver wrapper class at a later point, which
will choose an appropriate solver from the given matrix type from the list of
supported direct solver backends (LAPACK, UMFPACK, SuperLU, etc.).

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>