added UMFPACK as coarse grid solver for PPnD application
This commit enables the double-layered partitioning in the PartiDomainControl
for the PPnD application, which in turn enables the app to use UMFPACK as a
coarse-grid solver for the S-matrix multigrid solver. UMFPACK is now chosen
as the default coarse grid solver, if a) the build is configured with UMFPACK
support (obviously) and b) the '--level' command line option has been given
three parameters level-max, level-med and level-min. Typically you want to
use something like '--level n 0 0', to exploit the full hierarchy.

Also fixed a bunch of documentation errors and typos while at it.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>