implemeneted fully recursive PartiDomainControl (finally)
After almost 3 years of plans and preparations, this commit finally introduces
a long awaited milestone, which dramatically extends FEAT's list of parallel
FEM toolkit functionality: support for fully recursive partitioning!

In detail: The PartiDomainControl already had support for single- and double-
layered partitionings, but now it is also possible to create arbitrarily deep
multi-layered partitions, as long as some restrictions are honored.

The usage of the revamped PartiDomainControl is mostly unchanged, with the
one notable change that the second parameter of its contructor is now not
defaulted to 'false' (which indicates that the user's code does not support
multi- layered partitionings), so that every user has to specify this
parameter explicitly.

The desired levels are still specified via one of the overloads of the
'set_desired_levels()' function. One of these overloads accepts a String
object, which may be parsed from file or the command line. Please consult
the class documentation for details about the specification of levels.

All applications, which have been using the single- and/or double-layered
partitioning functionality of the PartiDomainControl before, should also work
with the new multi-layered version out-of-the-box, possibly with a few minor
code changes.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>