added Dist::Comm::null() + fixed comments + fixed Comm::[i]reduce
The documentation of the gather/scatter functions of Dist::Comm as well as
all of its related functions claimed that the recvcount and/or sendcount
parameters represented the total recv/send buffer size, which is not correct.
Instead, these counts represent the number of elements recieved from/send to
*each* process. Fixed that.

Moreover, MPI_IN_PLACE can be used for MPI_Reduce and MPI_Ireduce only on the
root process, but not for any other non-root process, even if both send and
recv buffers are given and are in fact identical.

Also added Dist::Comm::null(), which returns a new null communicator.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>