resolved quadratic runtime issues in RootMeshNode::extract_patch()
We have been using the PartiDomainControl and its "organ" RootMeshNode for
quite some time now without noticing that there have been two sub-algorithms
located in the RootMeshNode::extract_patch() function, which had a runtime
of O(N^2/P), where N is the number of elements in the to-be-partitioned base
mesh and P is the number of processes/partitions. As we usually have been
considering scenarious where N=O(P), this problem was not discovered until
Dirk started playing around with 3D base-meshes with more than 10^5 elements
and only wanting to have a handful of partitions. To my shame, I have to
admit that I am responsible for introducing both of these bottlenecks.

The first bottleneck was hidden within the computation of the ranks-at-rank
adjacency graph, which required the compositon of the elements-at-rank and
the vertices-at-element graphs as a first step. See the notes inside the
extract_patch() function for details about the problem and its solution.

The second bottleneck was hidden in the PatchHaloFactory and its helper class
PatchHaloBuild, whose build() member function also had a runtime of O(N^2/P).
This class has been reimplemented by me in commit 69ec85fece9c4cb8 roughly
18 months ago during a minor revisement of the MeshNode class.
Now the funny part is that the previous implementation had linear runtime,
so we are dealing with a perfect case of Verschlimmbesserung here. Therefore,
the previous imlementation was restored and the quadratic runtime problem
thus resolved.

As a possible minor drawback, the new implementation may require a slightly
more memory and may even have longer absolute runtimes in the N=O(P) case
that we are usually considering. However, these effects (if any) should be
negligible since the runtime is now guaranteed to be linear for all
practically relevant scenarious.

Note to self:
Always make sure that graph compositions have linear runtime for ALL relevant
scenarious before implementing them.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>