Suppress unitialised value warning in hypre
This commit introduces the file valgrind.supp, which contains
suppression directives for valgrind.

New suppressions can be created by running valgrind with the
--gen-suppressions=yes parameter.
These may than be added with a proper name to the valgrind.supp file.

In this special case the hypre thirdparty library fails to properly
initialise some dummy communicator stuff in its serial mode.

In addition, this commits lets valgrind be more verbose on the
parameters (especially the suppression files) used.

Signed-off-by: Dirk Ribbrock <dirk.ribbrock@math.tu-dortmund.de>