implemented streamline diffusion stabilisation in BurgersAssembler
This commit adds the Samarskij-style streamline diffusion, as suggested in
Turek's CFD book, to the scalar and blocked matrix assembly methods of our
BurgersAssembler class. For details on the usage of the stabilisation, please
check the documentation of the BurgersAssembler class.

Please note that although the BurgersAssembler class can assemble both the
streamline diffusion operator as well as the Frechet derivative of the
convective operator, it cannot assemble the Frechet derivative of the
streamline diffusion operator (yet), so in consequence one would lose the
quadratic convergence of Newton's method if streamline diffusion is enabled.

This commit also adds support for the streamline diffusion to the CCND app,
which can be enabled by specifying the stabilisation parameter using the
'--upsam <delta>' command line option.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>