reimplemented cell volume computation in standard trafo evaluator
This commit adds the 'volume' member function to the Trafo evaluator interface
as well as its standard implementation. To avoid conflicts, the implementation
of the CellVolumeEvaluator class has been removed and all its uses replaced by
corresponding calls of the new 'volume' function.

Internally, each of the 'volume' functions computes the Jacobian matrix
determinant and scales it by a corresponding factor to obtain the volume.
The only exception is the Hexahedron version, which uses a hard-coded 2x2x2
Gauss-Legendre cubature rule to compute the volume, because the barycentre
rule is insufficient in this case.

Moreover, this commit adds the cell volumes as an additional cell field to the
"mesh2vtk" tool output.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>