implemented directional mesh width approximation functions
This commit adds a new member function 'width_directed()' to the trafo
evaluator classes, which takes a (preferably) normalised direction vector
as an input argument and returns an approximation of the currently selected
cell's mesh width w.r.t. to the given direction.

The  internal implementation uses the follwing formula for the approximation:

                L * || J_T^{-1} * v ||_2^{-1}

where
* J_T is the Jacobian matrix of the transformation from the reference
  hypercube or a (equilateral) unit simplex evaluated in the barycentre
* v is the given input direction vector
* L is a scaling factor which ensures that the returned with is equal to 1
  for a unit element.

Note:
This function is required for the implementation of various stabilisations,
most notably the Samarskij-style streamline diffusion method.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>