moved debug tools to new 'area51' directory
This new 'area51' directory acts as a collection of all types of applications
and tools, which have been written to act as a development and testing ground
during the development and implementation phase of some complex new feature
and which may come in handy if that particular feature has to be enhanced,
reverified, bugfixed or somehow else modified in future.

Up to now, this type of application/tool was stored in the 'tools' directory,
usually prefixed by 'dbg_', which now all have been moved to the 'area51'.

This commit also adds two new debug apps, which have been used in the recent
implementation of the directed mesh width computation (dbg_mesh_width) and
the streamline diffusion (dbg_streamdiff).

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>