implemented macro-wise matrix-based version of Vanka smoother
This commit adds the Solver::AmaVanka class, which implements an additive
macro-wise Vanka smoother, which stores its operator as a matrix, so that
each application of this smoother boils down to a single matrix-vector
multiplication.

This class supports various matrix types, including SaddlePointMatrix and
TupleMatrix instances of SparseMatrixBCSR sub-blocks. In the case of the
SaddlePointMatrix, this new implementation is mathematically equivalent to
the smoother implemented in the "old" Solver::Vanka class of the type
VankaType::block_full_add.

Note that any other matrix type, one has to explicitly specify the adjacency
graphs representing the macro-dofs for each finite element space involved.

Also note that this smoother can in principle also be applied on Mem::CUDA,
however, currently there exists no generic implementation of the matrix-vector
product for SparseMatrixBCSR on Mem::CUDA, which is a show-stopper for this
case.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>