removed support for MPI-v2-only mode
This mode was added to ensure that FEAT could be compiled on the (already
decomissioned) JUQUEEN cluster. Any modern cluster should support MPI v3,
and because we want to make use of many of MPI's non-blocking features,
we simply drop the support for MPI v < 3 only clusters to avoid code forks.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>