This introduces two examples how to use the CGAL to analyse a mesh.
Both demoapplications obviously need to be compiled with CGAL.
Besides, both applications show how it is possible to obtain
information from a mesh on the absolute low-level and maybe do
some modifications.
One of the applications is parallel, the other one not.
The reason is that some things are better seen in the serial case
because it is a bit easier, but in a real-world-application
one will be more interested in the parallel case.

Signed-off-by: Malte Schuh <malte.schuh@mathematik.tu-dortmund.de>