Disable early tracking of variables in opt mode
Usually a variable is tracked from its origin to its final position
through all optimisation stages.
Some applications are too big for all variables to be tracked.
In this case gcc/clang ommit a warning, that this debugging support is
dropped and recompilation starts with early tracking disabled.
This commit disables early variable tracking in the first place for the
very big applications.

Signed-off-by: Dirk Ribbrock <dirk.ribbrock@math.tu-dortmund.de>