refactored and updated Spline chart to Bezier chart
This commit refactores and restructures the Spline chart, which is used to
parameterise composite Bezier curves for the boundary description in mesh
files.

There are 3 major changes in this commit:

Firstly, the class Geometry::Atlas::Spline has been renamed to
Geometry::Atlas::Bezier. Moreover, the corresponding XML tag has also been
renamed from Spline to Bezier.

Secondly, the encoding of the control and vertex points for the Bezier curves
in the mesh files has been changed. Previously, the control points were
defined in the same line as the start vertex of a curve segment, succeeding
the latter one. Now, the control points are in the same line as the end vertex
of a curve segment, preceeding the latter one. This corresponds to the
"stardard way" of describing Bezier curves, which conforms to the way that
most other apps and tools use.

Thirdly, this commit also introduces two additional culling checks for the
internal point-to-curve projection method, which is used for both the implicit
mesh parameterisation as well as for the signed distance computation, which is
required for mesh deformation aka r-adaptivity.

An update for the DeViSoR Grid3D exporter will follow shortly.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>