minor PropertyMap i/o function updates
This commit introduces three changes to the i/o functions of the PropertyMap
class:

For the sake of consistency, the input/output functions have been renamed
from 'parse' to 'read' and from 'dump' to 'write', respectively.

Moreover, overloads of the read/write functions accepting a communicator have
been added to enable efficient i/o for (massivley) parallel apps.

Finally, the functionality to include other INI files from another INI file
via the '@include' statement had to be removed, because this functionality
can not be ported (directly) to parallel applications. Note, however, that
this functonality has never been used in practice, so probably noone will
miss it.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>