unified memory
This eradicates any usage of Mem::main and Mem::cuda
It introduces an enum PreferredBackend which is used to select the
preferred backend (generic, mkl, cuda) at runtime.
It also includes a hierarchy of failovers to allow for mainly cuda
centric calculations with some fill ins from mkl or generic code, where
no cuda implementation is present.

Down the road, we got rid of a bunch of deprecated classes, including
but not restricted to:
  * SolverFactor
  * MatrixStock
  * SparseMatrixCOO
  * SParseMatrixELL
  * SPAI
  * AMG (self written, the BoomerAMG interface is still in use)

This commit is joined work of Peter Zajac, Lisa Walter and Dirk
Ribbrock.
It is squashed into one giant ball of mud to evade the sisyphean
undertaking of rebasing hundrets of commits against any change made in
develop in the mean time.

Signed-off-by: Dirk Ribbrock <dirk.ribbrock@math.tu-dortmund.de>