added basic support for the FloatX third-party library
This commit allows us to use "reduced precision" floating point types in FEAT,
which can be used to emulate half-precision. Unfortunately, the other library
that we wanted to use for half-precision, namely the "half" library, has been
developed in the age when mankind thought that expression templates are funky,
completely ignoring the fact that this concept completely flaks the "normal"
template code paradigms, which in consequence meant that using named library
was practically impossible in FEAT. Duh.

It is important to say that the FloatX library is "work-in-progress" and that
it is still a long way from a stable release. Also, the current implementation
interally must use a 'double' value for storage, so one does not save any
memory or bandwidth by using these reduced floating types. The primary goal
of that library is to emulate the reduced precision arithmetic to study its
numerical effects on larger scientific codes, which is also what we are
primarily interested in.

Also note that currently there is no support for MPI parallelisation of the
FloatX class templates via our DistIO interface, because the library offers
a class template, which can be configured at compile-time, rather than just
a single class. This can be added lateron, should the need arise.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>