minor fixes in Math functions + Math test
This commit introduces some updated stopping criterions for the Math::exp
and Math::log implementations, because the previous ones were too simple and
could lead to infinite loops when using low precision floating point classes,
especially ones offered by third-party libraries.

As an example, the sequence y_k of the well-known exponential series used in
the Math::exp function is strictly increasing for x > 0 and alternating for
x < 0, which can be combined into the most beautiful C++ crypto-expression:

            x > 0 ? yl < y : n & 1 ? y < yl : yl < y

Also nice, but not quite as fancy is the sequence y_k of the Newton iteration
used in the Math::log function, which is strictly increasing for x > 1 and
strictly decreasing for x < 1. You can combine that into an expression for
homework.

Finally, the MathTest has been modified so that the tolerances have to be
specified explicitly rather than being computed automatically. Reason being:
automatic computation of tolerances via Math::pow and Math::eps is fancy
as long as those two functions can be assumed to work correctly, which is,
in fact, exactly what the MathTest is trying to validate. Duh.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>