implemented multi-threaded DomainAssembler and corresponding Job/Task classes
This commit finally introduces a multi-threaded assembler for the various finite element related
matrix/vector assemblies based on (bi)linear forms, which has been on my todo-list for almost 10
years now. Quite similar to the already existing (and still somewhat provisional) TraceAssembler,
one needs to create a DomainAssembler object for each mesh/trafo object over which assemblies have
to be performed, so in the typical case one has one domain assembler for each trafo/mesh in a
multigrid hierarchy. Also, in analogy to the TraceAssembler, the DomainAssembler has to be
configured and "compiled" by calling the corresponding member functions before it can be used.

The DomainAssembler supports various threading strategies, which control how the synchronization
and element processing order are performed, most importantly the layered and colored strategies,
which are based on a Cuthill-McKee and Red-Black-Coloring scheme, respectively. Of course it is
also possible to assemble in a single-threaded manner as before, which is also the default setting,
i.e. the DomainAssembler will perform the assembly only on the calling thread by default unless
the user explicitly requests multithreading prior to compiling the assembler.

The multi-threading approach is implemented using the C++11 std::thread, std::condition_variable and
std::mutex classes, so in consequence it's not necessary to compile FEAT with OpenMP to use multiple
threads for the assembly. The reason for this is that the layered threading strategies require some
slightly complex thread synchronization techniques which, at least to my knowledge, are not
available in OpenMP.

The DomainAssembler does not directly assemble the already well-known BilinearOperator and
LinearFunctional interface implementations, but insteady intoduces a new duck-type interface
based on a "Job" interface class and its nestes "Task" class, which are documented in the
DomainAssemblyJob and DomainAssemblyJob::Task classes, respectively. This new Job/Task interface
allows for much more complex assemblers to be implemented in that same interface, even if the
actual assembler is more complex than a simple BilinearOperator implementation, such as e.g. was
the case with the BurgersAssembler, which now has been re-implemented in several BurgersAssemblyJob
classes. Nevertheless, it is of course still possible to assemble operators and functionals that
implement the original BilinearOperator and LinearFunctional interaces by using corresponding
wrapper classes or -- even more convenient -- helper functions, which take care of the wrapping.

The following assembly job classes have been implemented to superseed/wrap the corresponding classes:

* BilinearOperatorMatrixAssemblyJob1/2:
  Wrapper for BilinearOperator implementations, which assembles the operator into a matrix.
  This effectively reimplements the functionality of BilinearOperatorAssembler::assemble_matrix1/2.
  This wrapper class is used by the helper function Assembly::assemble_bilinear_operator_matrix_1/2.

* LinearFunctionalAssemblyJob:
  Wrapper for LinearFunctional implementations, which assembles the functional into a vector.
  This effectively reimplements the functionality of LinearFunctionalAssembler:::assemble_vector.
  This wrapper class is used by the helper function Assembly::assemble_linear_functional_vector.

* ForceFunctionalAssemblyJob:
  This is effectively a wrapper around a ForceFunctional implementation, which assembles a forces
  functional wrapper around an analytic function into a vector. This class is used by the helper
  function Assembly::assemble_force_function_vector, which takes an analytic function as input.

* Burgers{Blocked|Scalar}{Matrix|Vector}AssemblyJob:
  These classes (and their corresponding base-class hierarchies) reimplement the functionality of
  the BurgersAssembler and its various assembly member functions. This class is not a wrapper.

* AnalyticFunctionIntegralJob:
  This job class implements the functionality of integrating an analytic function (and possibly its
  derivatives) as well as computing its H0/H1/H2/L1/Lmax-(semi-)norms, which was not implemeneted
  before. This class is used by the helper function Assembly::integrate_analytic_function.

* DiscreteFunctionIntegralJob:
  This job class implements the functionality of integrating a discrete finite element function (and
  possibly its derivatives) as well as computing its H0/H1/H2/L1/Lmax-(semi-)norms, which was not
  implemeneted before. This class is used by the helper function Assembly::integrate_discrete_function.

* ErrorFunctionIntegralJob:
  This job class implements the functionality of integrating a (analytic - discrete) error function
  (and possibly its derivatives) as well as computing its H0/H1/H2/L1/Lmax-(semi-)norms. This class
  effectively superseeds the ScalarErrorComputer and VectorErrorComputer classes and it is used by
  the helper function Assembly::integrate_error_function.

All tutorials have been modified to use the new DomainAssembler class as well as the new helper
functions implemented in the assembly/domain_assembler_helpers.hpp header file. However, the apps
in the applications directory have not been adapted yet; this will be done later.

Also note that all previous assembler classes still exist and their functionality has not been
changed by this commit, so it is still safe to use them.

This commit also includes a ton of smaller changes like documentation fixes and some new member
functions which were required by the new assembler.

It is also important to mention that the threading strategies of the new DomainAssembler class are
partially based on the permutation strategies of the Geometry module; however, the are still some
optimizations left to be implemented here.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>