Replace inheritance requirement for checkpointable containers
This commit introduces a wrapper for checkpointable classes,
called CheckpointableWrapper.
Any class, that implements the methods, defined by the Checkpointable
interface, can be registered as checkpointable by using this wrapper
class.
This releases the burden from all containers to implement the
Checkpointable interface at all costs and replaces it by an opt-in
model.