added basic test app for FEM on sphere problems
This commit adds the first batch of "FEM on manifolds" functionality,
which enable us to solve basic PDEs on the 3D sphere. Besides the new
test app, this commit introduces a two new major classes:

* Geometry::UnitSphereFactory
A Factory class that can create a triangular or quadrilateral unit-sphere
mesh. For triangles, the generated mesh is an icosahedron, whereas for
quadrilaterals, it is simply a shifted and scaled unit-cube surface mesh.

* Trafo::IsoSphere::Mapping
An implementation of the Trafo interface that performs a higher-order
transformation from the 2D reference triangle/quadrilateral onto the
3D sphere. Currently, only the second-order quadratic transformation
is implemented; higher orders may be added in future. Note that one
can use the standard trafo mapping for a first order transformation.

This commit also adds the Assembly::Common::LaplaceBeltramiOperator,
which is used to assemble the Poisson-Problem on a sphere, as well as
the Analytic::Common::SphereSinBubbleFunction, which is a useful
reference solution for the Sphere-Poisson-Problem.

Furthermore, a few minor adjustments had to be made to make the code work.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>