Added cray compiler support
Because a cray compiler usually can only be found on cray machines,
the default compiler binary names are set to cc/CC.
These are set on cray machines via the programming environment system
and point to the correct compiler binaries.

This makes this compiler build id token a hybrid.
It can be used to choose a cray compiler, but moreover it shall be used
on cray systems to choose any compiler that is prechoosen by the
cray programming environment and called via the cc/CC aliases.

Signed-off-by: Dirk Ribbrock <dirk.ribbrock@math.tu-dortmund.de>