implemented 3D tetrahedral Lagrange-3 (aka P3) element
This commit adds the implementation of the 3D P3 element, which has been
waiting on the todo-list for more than two and a half years. In the end,
implementing this element was much less dramatic than expected.

It is worth mentioning that this element cannot be used in practice yet,
because the highest degree cubature rule that we have for tetrahedra right
now is of fifth degree -- but we require at least sixth degree for the
proper use of the P3 element. Once more tetrahedral cubature rules have
been implemented, this element will be ready to use.

This commit also introduces a simple jump-operator assembly to the trace
assmbler, because this was needed for debugging the element.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>