added cubature debug tool + more tetrahedral cubature rules
This commits adds a new debug tool, which can determine the
order of exactness of a cubature rule by explicitly integrating
all monomials and comparing the results to the analytic values.

Furthermore, this commit adds a set of tetrahedral cubature rules
based on an article by Lee Shunn and Frank Ham. These rules have
been chosen as the rules for the 'auto-degree' alias for tetrahedra,
therefore replacing the previous ones that were based on rules
from the famous cubature book of Stroud.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>