fix 2lvl-matrix assembly for StructuredMesh
The problem with the assembly routine was that the relationship between
the numberings of the coarse and the fine mesh elements was hardcoded
for ConformalMesh. The changes fix the assembly for all dimensions.

With the fix there comes a bit of restructuring: The RefinementAdjactor
has been moved from kernel/assembly/symbolic_assembler.hpp to
kernel/geometry/intern/coarse_fine_cell_mapping.hpp, is now called
CoarseFineCellMapping and is initialized differently (see doc).
This Adjactor is now also used in the assembly routines in
kernel/assembly/grid_transfer.hpp even though there is no need for a
full Adjactor. But it is cleaner to not split the class up.

Two other additions that come along with the fix:
* The RefineFactory for StrucutredMesh was missing and has been added
  therefore.
* The Mesh classes now have the bool indicator is_structured to make it
  easier to distinguish between structured and unstructured meshes.