proposed fix to ILU precond to make it work for triangular matrices.
According to cppreference.com std::vector::data() returns a
non-dereferenceable pointer if the vector is empty. (the pointer may or may
not be a nullptr).

This commit introduces checks for empty upper and lower triangular parts and
sets the corresponding pointers to nullptrs if those data vectors are empty.
Furthermore an additional check is performed in the factorization stage to
avoid dereferncing such a nullptr.

Signed-off-by: Jonas Duennebacke <jduenneb@math.tu-dortmund.de>