proposed ILU factorisation fix v1.1
Up to now, ILU behaved potentially erroneously then trying to factorise
nonsymmetric matrices, e.g. like the following 5x5 matrix:

  [ 1 . . . ]
  [ . 2 3 . ]
  [ 4 . 5 . ]
  [ . . . 6 ]

Here, when trying to eliminate the entry '4' by subtracting the first row,
the previous algorithm subtracted the entry '3' from the following row
from the main diagonal entry '5' due to a missing end-of-row check.

This missing end-of-row check was also the cause for the symptoms that
have been tackled by the previous commits.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>