added more meaningful syntax/eval error handling for Analytic::ParsedFunction
This commit adds a proper error message to the ParseError exception that is
thrown by the parse() method of the ParsedFunction class. Previously, the
error message was only 'Failed to parse function', whereas now one gets
something of the form:

ParseError: Syntax error: Unknown identifier
>>> 'x+sin(2*pi+3^k+y)'
>>>---------------^

Furthermore, this commit also adds check for errors during evalution, e.g.
to check for division by zero. A new ParsedFunctionEvalError exception class
has been added for this.

Finally, the ParsedFunction::Evaluator constructor now calls the ForceDeepCopy()
method of the internal parser to force a deep copy, which may be cruicial to
avoiding race conditions in multithreaded code.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>