added quick evaluation function for Analytic function implementations
This commits adds a bunch of small helper functions

* eval_value[_x]
* eval_gradient[_x]
* eval_hessian[_x]

to the Analytic namespace, which can be used to quickly evaluate an
implementation an Analytic::Function without the need of explicitly
creating evaluation traits and evaluators.

The functions without the '_x' suffix take a Tiny::Vector as the evaluation
point, where the '_x' suffixed functions take one, two or three coordinates
as parameters.

Be aware that these functions should only be used for testing and debugging
for the sake of efficiency.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>